Split open packs into separate files for components
