+++ packages/site/src/components/cards/PinCardToProfileButton.tsx
@@ -1,5 +1,5 @@
 import { USER_API } from '@/constants';
-import { Form, SubmitButton } from '../form';
+import { Form, SubmitButton } from '../form/Form';
 import { createSignal } from 'solid-js';
 
 export default function PinCardToProfileButton(props: {
