--- packages/site/src/components/form/index.tsx
@@ -1,1 +1,0 @@
-export * from './Form';
